Refactor : CHANGE CLASS CAR FILE TO MODEL
