+++ src/main/java/racingcar/Print.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.Model.Car;
+
 import java.util.List;
 
 public class Print {
